+++ src/main/java/codesAndStandards/springboot/registrationlogin/repository/RoleRepository.java
@@ -4,17 +4,12 @@ import codesAndStandards.springboot.registrationlogin.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-//public interface RoleRepository extends JpaRepository<Role, Long> {
-//    Role findByUsername(String username);
-//}
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
 
     Optional<Role> findByRoleName(String roleName);
-//    Optional<Role> findByUsername(String username);
 
     boolean existsByRoleName(String roleName);
-//    boolean findByUsername(String username);
 }
