+++ src/main/java/codesAndStandards/springboot/registrationlogin/service/DocumentService.java
@@ -1,35 +1,29 @@
 package codesAndStandards.springboot.registrationlogin.service;
 
 import codesAndStandards.springboot.registrationlogin.dto.DocumentDto;
-import codesAndStandards.springboot.registrationlogin.entity.Document;
 import org.springframework.web.multipart.MultipartFile;
+
 import java.util.List;
 
 public interface DocumentService {
-    // ✅ UPDATED: Added groupIds parameter
+
     void saveDocument(DocumentDto documentDto, MultipartFile file, String username, String groupIds) throws Exception;
 
     List<DocumentDto> findAllDocuments();
 
-    DocumentDto findDocumentById(Integer id);
+    DocumentDto findDocumentById(Long id);
 
-    // ✅ UPDATED: Added groupIds parameter
-    void updateDocument(Integer id, DocumentDto documentDto, MultipartFile file, String username, String groupIds) throws Exception;
+    void updateDocument(Long id, DocumentDto documentDto, MultipartFile file, String username, String groupIds) throws Exception;
 
-    void deleteDocument(Integer id);
+    void deleteDocument(Long id);
 
-    String getFilePath(Integer id);
+    String getFilePath(Long id);
 
     List<DocumentDto> findDocumentsAccessibleByUser(Long userId);
 
-    String getGroupNamesForDocument(Integer id);
-    /**
-     * ⭐ NEW - Check if user has access to document
-     */
-    boolean hasUserAccessToDocument(Long userId, int documentId);
-    /**
-     * ⭐ NEW - Get list of accessible document IDs for a user
-     * This uses the SAME logic as the document library
-     */
-    List<Integer> getAccessibleDocumentIds(Long userId);
+    String getGroupNamesForDocument(Long id);
+
+    boolean hasUserAccessToDocument(Long userId, Long documentId);
+
+    List<Long> getAccessibleDocumentIds(Long userId);
 }
